+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## 0.0
+- **2018-10.12**:
+    - Fixed bug which lead to uncontrolable divergence in the OpenCL integrator.
 - **2018-09-21**:
     - fixed bug in rk4 butcher schema of OpenCLKernel.
     - optimized OpenCLKernel by using local memory for jumping buffers.